+++ src/main/java/com/sbproject/piclocket/PicLocketApplication.java
@@ -1,11 +1,12 @@
 package com.sbproject.piclocket;
 
+import com.sbproject.piclocket.config.AwsProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.hibernate.autoconfigure.HibernateJpaAutoConfiguration;
-import org.springframework.boot.jdbc.autoconfigure.DataSourceAutoConfiguration;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
+@EnableConfigurationProperties(AwsProperties.class)
 public class PicLocketApplication {
 
     public static void main(String[] args) {